Model Reference

This page collects all information about the core components of IESopt, their properties (and default values), their internal mathematical formulations, and some examples of how to use them. See the API Reference for a complete list of all available functions and types.

IESopt.Connection — Type

A Connection is used to model arbitrary flows of energy between Nodes. It allows for limits, costs, delays, ...

Parameters

Name	Mandatory	Values	Default	Description

node_from	yes	string	-	This Connection models a flow from node_from to node_to (both are Nodes).

node_to	yes	string	-	This Connection models a flow from node_from to node_to (both are Nodes).

Detailed Model Reference

Variables

flow

How to?

Access this variable by using:

# Julia
component(model, "your_connection").var.flow

# Python
model.get_component("your_connection").var.flow

You can find the full implementation and all details here: IESopt.jl.

Add the variable representing the flow of this connection to the model. This can be accessed via connection.var.flow[t].

Additionally, the flow gets "injected" at the Nodes that the connection is connecting, resulting in

\[\begin{aligned} & \text{connection.node}_{from}\text{.injection}_t = \text{connection.node}_{from}\text{.injection}_t - \text{flow}_t, \qquad \forall t \in T \\ & \text{connection.node}_{to}\text{.injection}_t = \text{connection.node}_{to}\text{.injection}_t + \text{flow}_t, \qquad \forall t \in T \end{aligned}\]

math

For "PF controlled" Connections (ones that define the necessary power flow parameters), the flow variable may not be constructed (depending on specific power flow being used). The automatic result extraction will detect this and return the correct values either way. Accessing it manually can be done using connection.exp.pf_flow[t].

Expressions

pf_flow

How to?

Access this expression by using:

# Julia
component(model, "your_connection").exp.pf_flow

# Python
model.get_component("your_connection").exp.pf_flow

You can find the full implementation and all details here: IESopt.jl.

Construct the JuMP.AffExpr holding the PTDF based flow of this Connection.

This needs the global addon Powerflow with proper settings for mode, as well as properly configured power flow parameters for this Connection (pf_V, pf_I, pf_X, ...).

Constraints

flow_bounds

How to?

Access this constraint by using:

# Julia
component(model, "your_connection").con.flow_bounds

# Python
model.get_component("your_connection").con.flow_bounds

You can find the full implementation and all details here: IESopt.jl.

Add the constraint defining the bounds of the flow (related to connection) to the model.

Specifiying capacity will lead to symmetric bounds ($\text{lb} := -capacity$ and $\text{ub} := capacity$), while asymmetric bounds can be set by explicitly specifiying lb and ub.

Note

Usage of etdf is currently not fully tested, and not documented.

Upper and lower bounds can be "infinite" (by not setting them) resulting in the repective constraints not being added, and the flow variable therefore being (partially) unconstrained. Depending on the configuration the flow is calculated differently:

if connection.etdf is set, it is based on an ETDF sum flow,

if connection.exp.pf_flow is available, it equals this

else it equal connection.var.flow

This flow is then constrained:

\[\begin{aligned} & \text{flow}_t \geq \text{lb}, \qquad \forall t \in T \\ & \text{flow}_t \leq \text{ub}, \qquad \forall t \in T \end{aligned}\]

math

Constraint safety

The lower and upper bound constraint are subject to penalized slacks.

Objectives

cost

How to?

Access this objective by using:

# Julia
component(model, "your_connection").obj.cost

# Python
model.get_component("your_connection").obj.cost

You can find the full implementation and all details here: IESopt.jl.

Add the (potential) cost of this connection to the global objective function.

The connection.cost setting introduces a fixed cost of "transportation" to the flow of this Connection. It is based on the directed flow. This means that flows in the "opposite" direction will lead to negative costs:

\[\sum_{t \in T} \text{flow}_t \cdot \text{cost}_t \cdot \omega_t\]

math

Here $\omega_t$ is the weight of Snapshot t.

Costs for flows in both directions

If you need to apply a cost term to the absolute value of the flow, consider splitting the Connection into two different ones, in opposing directions, and including lb = 0.

IESopt.Decision — Type

A Decision represents a basic decision variable in the model that can be used as input for various other core component's settings, as well as have associated costs.

Parameters

Name	Mandatory	Values	Default	Description

lb	no	numeric	0	Minimum size of the decision (considered for each "unit" if count allows multiple "units").

Detailed Model Reference

Variables

fixed

How to?

Access this variable by using:

# Julia
component(model, "your_decision").var.fixed

# Python
model.get_component("your_decision").var.fixed

You can find the full implementation and all details here: IESopt.jl.

to be added

sos

How to?

Access this variable by using:

# Julia
component(model, "your_decision").var.sos

# Python
model.get_component("your_decision").var.sos

You can find the full implementation and all details here: IESopt.jl.

to be added

value

How to?

Access this variable by using:

# Julia
component(model, "your_decision").var.value

# Python
model.get_component("your_decision").var.value

You can find the full implementation and all details here: IESopt.jl.

Add the variable describing the value of this decision to the model. If lower and upper bounds (decision.lb and decision.ub) are the same, the variable will immediately be fixed to that value. This can be accessed via decision.var.value.

Expressions

Constraints

fixed

How to?

Access this constraint by using:

# Julia
component(model, "your_decision").con.fixed

# Python
model.get_component("your_decision").con.fixed

You can find the full implementation and all details here: IESopt.jl.

to be added

sos1

How to?

Access this constraint by using:

# Julia
component(model, "your_decision").con.sos1

# Python
model.get_component("your_decision").con.sos1

You can find the full implementation and all details here: IESopt.jl.

to be added

sos2

How to?

Access this constraint by using:

# Julia
component(model, "your_decision").con.sos2

# Python
model.get_component("your_decision").con.sos2

You can find the full implementation and all details here: IESopt.jl.

to be added

sos_value

How to?

Access this constraint by using:

# Julia
component(model, "your_decision").con.sos_value

# Python
model.get_component("your_decision").con.sos_value

You can find the full implementation and all details here: IESopt.jl.

to be added

Objectives

fixed

How to?

Access this objective by using:

# Julia
component(model, "your_decision").obj.fixed

# Python
model.get_component("your_decision").obj.fixed

You can find the full implementation and all details here: IESopt.jl.

sos

How to?

Access this objective by using:

# Julia
component(model, "your_decision").obj.sos

# Python
model.get_component("your_decision").obj.sos

You can find the full implementation and all details here: IESopt.jl.

Add the cost defined by the SOS-based value of this Decision to the model.

value

How to?

Access this objective by using:

# Julia
component(model, "your_decision").obj.value

# Python
model.get_component("your_decision").obj.value

You can find the full implementation and all details here: IESopt.jl.

Add the cost defined by the value of this Decision to the model:

\[\text{value} \cdot \text{cost}\]

math

IESopt.Node — Type

A Node represents a basic intersection/hub for energy flows. This can for example be some sort of bus (for electrical systems). It enforces a nodal balance equation (= "energy that flows into it must flow out") for every Snapshot. Enabling the internal state of the Node allows it to act as energy storage, modifying the nodal balance equation. This allows using Nodes for various storage tasks (like batteries, hydro reservoirs, heat storages, ...).

Basic Examples

A Node that represents an electrical bus:

bus:
  type: Node
  carrier: electricity

A Node that represents a simplified hydrogen storage:

store:
  type: Node
  carrier: hydrogen
  has_state: true
  state_lb: 0
  state_ub: 50

Parameters

Name	Mandatory	Values	Default	Description

carrier	yes	string	-	Carrier of this Node. All connecting components need to respect that.

has_state	no	true or false	false	If true, the Node is considered to have an internal state ("stateful Node"). This allows it to act as energy storage. Connect Connections or Units to it, acting as charger/discharger.

state_lb	no	numeric	$-\infty$	Lower bound of the internal state, requires has_state = true.

state_ub	no	numeric	$+\infty$	Upper bound of the internal state, requires has_state = true.

state_cyclic	no	eq, geq, or disabled	eq	Controls how the state considers the boundary between last and first Snapshot. disabled disables cyclic behaviour of the state (see also state_initial), eq leads to the state at the end of the year being the initial state at the beginning of the year, while geq does the same while allowing the end-of-year state to be higher (= "allowing to destroy energy at the end of the year").

state_initial	no	numeric	-	Sets the initial state. Must be used in combination with state_cyclic = disabled.

state_final	no	numeric	-	Sets the final state. Must be used in combination with state_cyclic = disabled.

state_percentage_loss	no	numeric	0	Per Snapshot percentage loss of state (loosing 1% should be set as 0.01).

nodal_balance	no	enforce, destroy, or create	enforce	Can only be used for has_state = false. enforce forces total injections to always be zero (similar to Kirchhoff's current law), create allows "supply < demand", destroy allows "supply > demand", at this Node.

sum_window_size	no	integer	-	TODO.

sum_window_step	no	integer	1	TODO.

Detailed Model Reference

Variables

pf_theta

How to?

Access this variable by using:

# Julia
component(model, "your_node").var.pf_theta

# Python
model.get_component("your_node").var.pf_theta

You can find the full implementation and all details here: IESopt.jl.

state

How to?

Access this variable by using:

# Julia
component(model, "your_node").var.state

# Python
model.get_component("your_node").var.state

You can find the full implementation and all details here: IESopt.jl.

Add the variable representing the state of this node to the model, if node.has_state == true. This can be accessed via node.var.state[t].

Additionally, if the state's initial value is specified via state_initial the following gets added:

\[\text{state}_1 = \text{state}_{initial}\]

math

Expressions

injection

How to?

Access this expression by using:

# Julia
component(model, "your_node").exp.injection

# Python
model.get_component("your_node").exp.injection

You can find the full implementation and all details here: IESopt.jl.

Add an empty (JuMP.AffExpr(0)) expression to the node that keeps track of feed-in and withdrawal of energy.

This constructs the expression $\text{injection}_t, \forall t \in T$ that is utilized in node.con.nodalbalance. Core components (Connections, Profiles, and Units) that feed energy into this node add to it, all others subtract from it. A stateless node forces this nodal balance to always equal 0 which essentially describes "generation = demand".

Constraints

last_state

How to?

Access this constraint by using:

# Julia
component(model, "your_node").con.last_state

# Python
model.get_component("your_node").con.last_state

You can find the full implementation and all details here: IESopt.jl.

Add the constraint defining the bounds of the node's state during the last Snapshot to the model, if node.has_state == true.

This is necessary since it could otherwise happen, that the state following the last Snapshot is actually not feasible (e.g. we could charge a storage by more than it's state allows for). The equations are based on the construction of the overall state variable.

\[\begin{aligned} & \text{state}_{end} \cdot \text{factor}^\omega_t + \text{injection}_{end} \cdot \omega_t \geq \text{state}_{lb} \\ & \text{state}_{end} \cdot \text{factor}^\omega_t + \text{injection}_{end} \cdot \omega_t \leq \text{state}_{ub} \end{aligned}\]

math

Here $\omega_t$ is the weight of Snapshot t, and $\text{factor}$ is either 1.0 (if there are now percentage losses configured), or (1.0 - node.state_percentage_loss) otherwise.

Constraint safety

The lower and upper bound constraint are subject to penalized slacks.

nodalbalance

How to?

Access this constraint by using:

# Julia
component(model, "your_node").con.nodalbalance

# Python
model.get_component("your_node").con.nodalbalance

You can find the full implementation and all details here: IESopt.jl.

Add the constraint describing the nodal balance to the model.

Depending on whether the node is stateful or not, this constructs different representations:

if node.has_state == true

\[\begin{aligned} & \text{state}_t = \text{state}_{t-1} \cdot \text{factor}^\omega_{t-1} + \text{injection}_{t-1} \cdot \omega_{t-1}, \qquad \forall t \in T \setminus \{1\} \\ \\ & \text{state}_1 = \text{state}_{end} \cdot \text{factor}^\omega_{end} + \text{injection}_{end} \cdot \omega_{end} \end{aligned}\]

math

Here $\omega_t$ is the weight of Snapshot t, and $\text{factor}$ is either 1.0 (if there are now percentage losses configured), or (1.0 - node.state_percentage_loss) otherwise. $\text{injection}_{t}$ describes the overall injection (all feed-ins minus all withdrawals). $end$ indicates the last snapshot in $T$. Depending on the setting of state_cyclic the second constraint is written as $=$ ("eq") or $\leq$ ("leq"). The latter allows the destruction of excess energy at the end of the total time period to help with feasibility.

if node.has_state == false

\[\begin{aligned} & \text{injection}_{t} = 0, \qquad \forall t \in T \\ \end{aligned}\]

math

This equation can further be configured using the nodal_balance parameter, which accepts enforce (resulting in $=$), create (resulting in $\leq$; allowing the creation of energy - or "negative injections"), and destroy ( resulting in $\geq$; allowing the destruction of energy - or "positive injections"). This can be used to model some form of energy that can either be sold (using a destroy Profile connected to this Node), or "wasted into the air" using the destroy setting of this Node.

state_bounds

How to?

Access this constraint by using:

# Julia
component(model, "your_node").con.state_bounds

# Python
model.get_component("your_node").con.state_bounds

You can find the full implementation and all details here: IESopt.jl.

Add the constraint defining the bounds of the node's state to the model, if node.has_state == true.

\[\begin{aligned} & \text{state}_t \geq \text{state}_{lb}, \qquad \forall t \in T \\ & \text{state}_t \leq \text{state}_{ub}, \qquad \forall t \in T \end{aligned}\]

math

Constraint safety

The lower and upper bound constraint are subject to penalized slacks.

Objectives

IESopt.Profile — Type

A Profile allows representing "model boundaries" - parts of initial problem that are not endogenously modelled - with a support for time series data. Examples are hydro reservoir inflows, electricity demand, importing gas, and so on. Besides modelling fixed profiles, they also allow different ways to modify the value endogenously.

Parameters

Name	Mandatory	Values	Default	Description

carrier	yes	string	-	Carrier of this Profile.

Detailed Model Reference

Variables

aux_value

How to?

Access this variable by using:

# Julia
component(model, "your_profile").var.aux_value

# Python
model.get_component("your_profile").var.aux_value

You can find the full implementation and all details here: IESopt.jl.

Add the variable that is used in this Profiles value to the model.

The variable var_value[t] is constructed and is linked to the correct Nodes. There are different ways, IESopt interprets this, based on the setting of profile.mode:

fixed: The value is already handled by the constant term of profile.exp.value and NO variable is constructed.

create, destroy, or ranged: This models the creation or destruction of energy - used mainly to represent model boundaries, and energy that comes into the model or leaves the model's scope. It is however important that create should mostly be used feeding into a Node (profile.node_from = nothing) and destroy withdrawing from a Node (profile.node_to = nothing). If lb and ub are defined, ranged can be used that allows a more detailled control over the Profile, specifying upper and lower bounds for every Snapshot. See _profile_con_value_bounds!(profile::Profile) for details on the specific bounds for each case.

This variable is added to the profile.exp.value. Additionally, the energy (that profile.exp.value represents) gets "injected" at the Nodes that the profile is connected to, resulting in

\[\begin{aligned} & \text{profile.node}_{from}\text{.injection}_t = \text{profile.node}_{from}\text{.injection}_t - \text{value}_t, \qquad \forall t \in T \\ & \text{profile.node}_{to}\text{.injection}_t = \text{profile.node}_{to}\text{.injection}_t + \text{value}_t, \qquad \forall t \in T \end{aligned}\]

math

Expressions

value

How to?

Access this expression by using:

# Julia
component(model, "your_profile").exp.value

# Python
model.get_component("your_profile").exp.value

You can find the full implementation and all details here: IESopt.jl.

Cosntruct the JuMP.AffExpr that keeps the total value of this Profile for each Snapshot.

This is skipped if the value of this Profile is handled by an Expression. Otherwise it is intialized based on profile.value.

Constraints

value_bounds

How to?

Access this constraint by using:

# Julia
component(model, "your_profile").con.value_bounds

# Python
model.get_component("your_profile").con.value_bounds

You can find the full implementation and all details here: IESopt.jl.

Add the constraint defining the bounds of this profile to the model.

This heavily depends on the mode setting, as it does nothing if the mode is set to fixed, or the value is actually controlled by an Expression. The variable can be accessed via profile.var.aux_value[t], but using the normal result extraction is recommended, since that properly handles the profile.exp.value instead.

Otherwise:

if profile.mode === :create or profile.mode === :destroy

\[\begin{aligned} & \text{aux_value}_t \geq 0, \qquad \forall t \in T \end{aligned}\]

math

if profile.mode === :ranged

\[\begin{aligned} & \text{value}_t \geq \text{lb}_t, \qquad \forall t \in T \\ & \text{value}_t \leq \text{ub}_t, \qquad \forall t \in T \end{aligned}\]

math

Here, lb and ub can be left empty, which drops the respective constraint.

Objectives

cost

How to?

Access this objective by using:

# Julia
component(model, "your_profile").obj.cost

# Python
model.get_component("your_profile").obj.cost

You can find the full implementation and all details here: IESopt.jl.

Add the (potential) cost of this Profile to the global objective function.

The profile.cost setting specifies a potential cost for the creation ("resource costs", i.e. importing gas into the model) or destruction ("penalties", i.e. costs linked to the emission of CO2). It can have a unique value for every Snapshot, i.e. allowing to model a time-varying gas price throughout the year.

The contribution to the global objective function is as follows:

\[\sum_{t\in T} \text{value}_t \cdot \text{profile.cost}_t \cdot \omega_t\]

math

Here $\omega_t$ is the weight of Snapshot t, and $\text{value}_t$ actually refers to the value of profile.exp.value[t] (and not only on the maybe non-existing variable).

IESopt.Unit — Type

A Unit allows transforming one (or many) forms of energy into another one (or many), given some constraints and costs.

Parameters

Name	Mandatory	Values	Default	Description

conversion	yes	string	-	The conversion expression describing how this Unit transforms energy. Specified in the form of "$\alpha \cdot carrier_1 + \beta \cdot carrier_2$ -> $\gamma \cdot carrier_3 + \delta \cdot carrier_4$". Coefficients allow simple numerical calculations, but are not allowed to include spaces (so e.g. (1.0/9.0) is valid). Coefficients are allowed to be NumericalInputs, resulting in column@data_file being a valid coefficient (this can be used e.g. for time-varying COPs of heatpumps).

capacity	yes	string	-	Maximum capacity of this Unit, to be given in the format X in/out:carrier where X is the amount, in or out (followed by :) specifies whether the limit is to be placed on the in- our output of this Unit, and carrier specifies the respective Carrier. Example: 100 in:electricity (to limit the "input rating").

Detailed Model Reference

Variables

conversion

How to?

Access this variable by using:

# Julia
component(model, "your_unit").var.conversion

# Python
model.get_component("your_unit").var.conversion

You can find the full implementation and all details here: IESopt.jl.

Add the variable describing the unit's conversion to the model.

This can be accessed via unit.var.conversion[t]; this does not describe the full output of the Unit since that maybe also include fixed generation based on the ison variable.

Info

This applies some heavy recalculation of efficiencies to account for minimum load and so on, that are currently not fully documented. This essentially comes down to the following: As long as minimum load is not enabled, that is rather simple (using the conversion expression to withdraw energy from the inputs and push energy into the outputs). If a separate minimum load conversion is specified it results in the following: (1) if running at minimum load the supplied minimum load conversion will be used; (2) if running at maximum capacity the "normal" conversion expression will be used; (3) for any point in-between a linear interpolation scales up all coefficients of the conversion expression to "connect" case (1) and (2).

conversion_connect

No documentation found.

IESopt._unit_var_conversion_connect! is a Function.

How to?

Access this variable by using:

# Julia
component(model, "your_unit").var.conversion_connect

# Python
model.get_component("your_unit").var.conversion_connect

You can find the full implementation and all details here: IESopt.jl.

ison

How to?

Access this variable by using:

# Julia
component(model, "your_unit").var.ison

# Python
model.get_component("your_unit").var.ison

You can find the full implementation and all details here: IESopt.jl.

Add the variable describing the current "online" state of the unit to the model.

The variable can be further parameterized using the unit.unit_commitment setting ("linear", "binary", "integer"). It will automatically enforce the constraints $0 \leq \text{ison} \leq \text{unitcount}$, with $\text{unitcount}$ describing the number of units that are aggregated in this unit (set by unit.unit_count). This can be accessed via unit.var.ison[t].

ramp

How to?

Access this variable by using:

# Julia
component(model, "your_unit").var.ramp

# Python
model.get_component("your_unit").var.ramp

You can find the full implementation and all details here: IESopt.jl.

Add the variable describing the per-snapshot ramping to the model.

This adds two variables per snapshot to the model (if the respective setting unit.enable_ramp_up or unit.enable_ramp_down is activated). Both are preconstructed with a fixed lower bound of 0. This describes the amount of change in conversion that occurs during the current snapshot. These can be accessed via unit.var.ramp_up[t] and unit.var.ramp_down[t].

These variables are only used for ramping costs. The limits are enforced directly on the conversion, which means this variable only exists if costs are specified!

startup

How to?

Access this variable by using:

# Julia
component(model, "your_unit").var.startup

# Python
model.get_component("your_unit").var.startup

You can find the full implementation and all details here: IESopt.jl.

Add the variable describing the per-snapshot startup to the model.

This adds a variable per snapshot to the model (if the respective setting unit.unit_commitment is activated). The variable can be further parameterized using the unit.unit_commitment setting ("linear", "binary", "integer"). It will automatically enforce the constraints $0 \leq \text{startup} \leq \text{unitcount}$, with $\text{unitcount}$ describing the number of units that are aggregated in this unit (set by unit.unit_count). This describes the startup that happens during the current snapshot and can be accessed via unit.var.startup.

Expressions

Constraints

conversion_bounds

How to?

Access this constraint by using:

# Julia
component(model, "your_unit").con.conversion_bounds

# Python
model.get_component("your_unit").con.conversion_bounds

You can find the full implementation and all details here: IESopt.jl.

Add the constraint defining the unit's conversion bounds to the model.

This makes use of the current min_capacity (describing the lower limit of conversion; either 0 if no minimum load applies or the respective value of the minimum load) as well as the online_capacity (that can either be the full capacity if unit commitment is disabled, or the amount that is currently active).

Depending on how the "availability" of this unit is handled it constructs the following constraints:

if !isnothing(unit.availability)

\[\begin{aligned} & \text{conversion}_t \geq \text{capacity}_{\text{min}, t}, \qquad \forall t \in T \\ & \text{conversion}_t \leq \text{capacity}_{\text{online}, t}, \qquad \forall t \in T \\ & \text{conversion}_t \leq \text{availability}_t, \qquad \forall t \in T \end{aligned}\]

math

This effectively results in $\text{conversion}_t \leq \min(\text{capacity}_{\text{online}, t}, \text{availability}_t)$.

if !isnothing(unit.availability_factor)

\[\begin{aligned} & \text{conversion}_t \geq \text{capacity}_{\text{min}, t}, \qquad \forall t \in T \\ & \text{conversion}_t \leq \text{capacity}_{\text{online}, t} \cdot \text{availability}_{\text{factor}, t}, \qquad \forall t \in T \end{aligned}\]

math

Info

If one is able to choose between using availability or availability_factor (e.g. for restricting available capacity during a planned revision to half the units capacity), enabling availability_factor (in this example 0.5) will result in a faster model (build and probably solve) since it makes use of one less constraint.

If no kind of availability limiting takes place, the following bounds are enforced:

\[\begin{aligned} & \text{conversion}_t \geq \text{capacity}_{\text{min}, t}, \qquad \forall t \in T \\ & \text{conversion}_t \leq \text{capacity}_{\text{online}, t}, \qquad \forall t \in T \end{aligned}\]

math

ison

How to?

Access this constraint by using:

# Julia
component(model, "your_unit").con.ison

# Python
model.get_component("your_unit").con.ison

You can find the full implementation and all details here: IESopt.jl.

Construct the upper bound for var_ison, based on unit.unit_count, if it is handled by an external Decision.

min_onoff_time

How to?

Access this constraint by using:

# Julia
component(model, "your_unit").con.min_onoff_time

# Python
model.get_component("your_unit").con.min_onoff_time

You can find the full implementation and all details here: IESopt.jl.

Add the constraints modeling min on- or off-time of a Unit to the model.

This constructs the constraints

\[\begin{align} & \sum_{t' = t}^{t + \text{min\_on\_time}} ison_{t'} >= \text{min\_on\_time} \cdot (ison_t - ison_{t-1}) \qquad \forall t \in T \\ & \sum_{t' = t}^{t + \text{min\_off\_time}} (1 - ison_{t'}) >= \text{min\_off\_time} \cdot (ison_{t-1} - ison_t) \qquad \forall t \in T \end{align} respecting `on_time_before` and `off_time_before`, and `is_on_before`. See the code for more details. !!! info "Aggregated units" This is currently not fully adapted to account for `Unit`s with `unit_count > 1`.\]

math

ramp

How to?

Access this constraint by using:

# Julia
component(model, "your_unit").con.ramp

# Python
model.get_component("your_unit").con.ramp

You can find the full implementation and all details here: IESopt.jl.

Add the auxiliary constraint that enables calculation of per snapshot ramping to the model.

Depending on whether ramps are enabled, none, one, or both of the following constraints are constructed:

\[\begin{aligned} & \text{ramp}_{\text{up}, t} \geq \text{conversion}_{t} - \text{conversion}_{t-1}, \qquad \forall t \in T \\ & \text{ramp}_{\text{down}, t} \geq \text{conversion}_{t-1} - \text{conversion}_{t}, \qquad \forall t \in T \end{aligned}\]

math

This calculates the ramping that happens from the PREVIOUS snapshot to this one. That means that if:

out[5] = 100 and out[4] = 50, then ramp_up[5] = 50 and ramp_down[5] = 0

ramp_up[1] = ramp_down[1] = 0

Info

This currently does not support pre-setting the initial states of the unit (it can be done manually but there is no exposed parameter), which will be implemented in the future to allow for easy / correct rolling optimization runs.

ramp_limit

How to?

Access this constraint by using:

# Julia
component(model, "your_unit").con.ramp_limit

# Python
model.get_component("your_unit").con.ramp_limit

You can find the full implementation and all details here: IESopt.jl.

Add the constraint describing the ramping limits of this unit to the model.

This makes use of the maximum capacity of the unit, which is just the total installed capacity. Both, up- and downwards ramps can be enabled separately (via unit.ramp_up_limit and unit.ramp_down_limit), resulting in either or both of:

\[\begin{aligned} & \text{ramp}_{\text{up}, t} \leq \text{ramplimit}_\text{up} \cdot \text{capacity}_\text{max} \cdot \omega_t, \qquad \forall t \in T \\ & \text{ramp}_{\text{down}, t} \leq \text{ramplimit}_\text{down} \cdot \text{capacity}_\text{max} \cdot \omega_t, \qquad \forall t \in T \end{aligned}\]

math

This does not make use of the ramping variable (that is only used for costs - if there are costs).

This calculates the ramping that happens from the PREVIOUS snapshot to this one. That means that if:

out[5] = 100 and out[4] = 50, then ramp_up[5] = 50 and ramp_down[5] = 0

ramp_up[1] = ramp_down[1] = 0

startup

How to?

Access this constraint by using:

# Julia
component(model, "your_unit").con.startup

# Python
model.get_component("your_unit").con.startup

You can find the full implementation and all details here: IESopt.jl.

Add the auxiliary constraint that enables calculation of per snapshot startup to the model.

Depending on whether startup handling is enabled, the following constraint is constructed:

\[\begin{aligned} & \text{startup}_{\text{up}, t} \geq \text{ison}_{t} - \text{ison}_{t-1}, \qquad \forall t \in T \end{aligned}\]

math

This calculates the startup that happens from the PREVIOUS snapshot to this one. That means that if:

ison[5] = 1 and ison[4] = 0, then startup[5] = 1

Objectives

marginal_cost

How to?

Access this objective by using:

# Julia
component(model, "your_unit").obj.marginal_cost

# Python
model.get_component("your_unit").obj.marginal_cost

You can find the full implementation and all details here: IESopt.jl.

Add the (potential) cost of this unit's conversion (unit.marginal_cost) to the global objective function.

\[\sum_{t \in T} \text{conversion}_t \cdot \text{marginalcost}_t \cdot \omega_t\]

math

ramp_cost

How to?

Access this objective by using:

# Julia
component(model, "your_unit").obj.ramp_cost

# Python
model.get_component("your_unit").obj.ramp_cost

You can find the full implementation and all details here: IESopt.jl.

Add the (potential) cost of this unit's ramping to the global objective function.

To allow for finer control, costs of up- and downwards ramping can be specified separately (using unit.ramp_up_cost and unit.ramp_down_cost):

\[\sum_{t \in T} \text{ramp}_{\text{up}, t} \cdot \text{rampcost}_{\text{up}} + \text{ramp}_{\text{down}, t} \cdot \text{rampcost}_{\text{down}}\]

math

startup_cost

How to?

Access this objective by using:

# Julia
component(model, "your_unit").obj.startup_cost

# Python
model.get_component("your_unit").obj.startup_cost

You can find the full implementation and all details here: IESopt.jl.

Add the (potential) cost of this unit's startup behaviour (configured by unit.startup_cost if unit.unit_commitment != :off).

\[\sum_{t \in T} \text{startup}_t \cdot \text{startupcost}\]

math