Model formulation

This page collects all information about the core components of IESopt, their properties (and default values), their internal mathematical formulations, and some examples of how to use them. See the API for a complete list of all available functions and types.

IESopt.Connection — Type

A Connection is used to model arbitrary flows of energy between Nodes. It allows for limits, costs, delays, ...

Parameters

Name	Mandatory	Values	Unit	Default	Description

node_from	yes	string	-	-	This Connection models a flow from node_from to node_to (both are Nodes).

node_to	yes	string	-	-	This Connection models a flow from node_from to node_to (both are Nodes).

carrier	no	string	-	-	Carrier of this Connection. If not given, automatically picks the carrier of the Nodes it connects. This parameter is not necessary, and only exists to allow for a more explicit definition.

capacity	no	numeric, col@file, decision:value	power	$+\infty$	The symmetric bound on this Connection's flow. Results in lb = -capacity and ub = capacity. Must not be specified if lb, ub, or both are explicitly stated.

lb	no	numeric, col@file, decision:value	power	$-\infty$	Lower bound of this Connection's flow.

ub	no	numeric, col@file, decision:value	power	$+\infty$	Upper bound of this Connection's flow.

cost	no	numeric	monetary (per energy)	-	Cost of every unit of energy flow over this connection that is added to the model's objective function. Keep in mind that negative flows will induce negative costs, which can be used to model revenues. Further, a bidirectional Connection (if lb < 0, which is the default, or if capacity is used) with a positive cost will lead to negative costs for the reverse flow. If you do not want this, split the Connection into two separate ones, each being unidirectional (with lb: 0). Remember, that these can share the same "capacity" (which is then set asub), even when using decision:value or col@file as value.

loss	no	$\in [0, 1]$	-	0	Fractional loss when transfering energy. This loss occurs "at the destination", which means that for a loss of 5%, set as loss: 0.05, and considering a Snapshot where the Connection has a flow value of 100, it will "extract" 100 from node_from and "inject" 95 into node_to. Since the flow variable is given as power, this would, e.g., translate to consuming 200 units of energy at node_from and injecting 190 units at node_to, if the Snapshot duration is 2 hours.

Detailed Model Reference

Variables

flow

How to?

Access this variable by using:

# Julia
component(model, "your_connection").var.flow

# Python
model.get_component("your_connection").var.flow

You can find the full implementation and all details here: IESopt.jl.

Add the variable representing the flow of this connection to the model. This can be accessed via connection.var.flow[t].

Additionally, the flow gets "injected" at the Nodes that the connection is connecting, resulting in

\[\begin{aligned} & \text{connection.node}_{from}\text{.injection}_t = \text{connection.node}_{from}\text{.injection}_t - \text{flow}_t, \qquad \forall t \in T \\ & \text{connection.node}_{to}\text{.injection}_t = \text{connection.node}_{to}\text{.injection}_t + \text{flow}_t, \qquad \forall t \in T \end{aligned}\]

math

For "PF controlled" Connections (ones that define the necessary power flow parameters), the flow variable may not be constructed (depending on specific power flow being used). The automatic result extraction will detect this and return the correct values either way. Accessing it manually can be done using connection.exp.pf_flow[t].

Expressions

pf_flow

How to?

Access this expression by using:

# Julia
component(model, "your_connection").exp.pf_flow

# Python
model.get_component("your_connection").exp.pf_flow

You can find the full implementation and all details here: IESopt.jl.

Construct the JuMP.AffExpr holding the PTDF based flow of this Connection.

This needs the global addon Powerflow with proper settings for mode, as well as properly configured power flow parameters for this Connection (pf_V, pf_I, pf_X, ...).

Constraints

flow_bounds

How to?

Access this constraint by using:

# Julia
component(model, "your_connection").con.flow_bounds

# Python
model.get_component("your_connection").con.flow_bounds

You can find the full implementation and all details here: IESopt.jl.

Add the constraint defining the bounds of the flow (related to connection) to the model.

Specifiying capacity will lead to symmetric bounds ($\text{lb} := -capacity$ and $\text{ub} := capacity$), while asymmetric bounds can be set by explicitly specifiying lb and ub.

Note

Usage of etdf is currently not fully tested, and not documented.

Upper and lower bounds can be "infinite" (by not setting them) resulting in the repective constraints not being added, and the flow variable therefore being (partially) unconstrained. Depending on the configuration the flow is calculated differently:

if connection.etdf is set, it is based on an ETDF sum flow,

if connection.exp.pf_flow is available, it equals this

else it equal connection.var.flow

This flow is then constrained:

\[\begin{aligned} & \text{flow}_t \geq \text{lb}, \qquad \forall t \in T \\ & \text{flow}_t \leq \text{ub}, \qquad \forall t \in T \end{aligned}\]

math

Constraint safety

The lower and upper bound constraint are subject to penalized slacks.

Objectives

cost

How to?

Access this objective by using:

# Julia
component(model, "your_connection").obj.cost

# Python
model.get_component("your_connection").obj.cost

You can find the full implementation and all details here: IESopt.jl.

Add the (potential) cost of this connection to the global objective function.

The connection.cost setting introduces a fixed cost of "transportation" to the flow of this Connection. It is based on the directed flow. This means that flows in the "opposite" direction will lead to negative costs:

\[\sum_{t \in T} \text{flow}_t \cdot \text{cost}_t \cdot \omega_t\]

math

Here $\omega_t$ is the weight of Snapshot t.

Costs for flows in both directions

If you need to apply a cost term to the absolute value of the flow, consider splitting the Connection into two different ones, in opposing directions, and including lb = 0.

IESopt.Decision — Type

A Decision represents a basic decision variable in the model that can be used as input for various other core component's settings, as well as have associated costs.

Parameters

Name	Mandatory	Values	Unit	Default	Description

lb	no	numeric	-	0	Minimum size of the decision value (considered for each "unit" if count allows multiple "units").

ub	no	numeric	-	$+\infty$	Maximum size of the decision value (considered for each "unit" if count allows multiple "units").

cost	no	numeric	monetary (per value)	0	Cost that the decision value induces, given as $cost \cdot value$.

fixed_value	no	numeric	-	-	If mode: fixed, this value is used as the fixed value of the decision. This can be useful if this Decision was used in a previous optimization and its value should be fixed to that value in the next optimization (applying it where ever it is used, instead of needing to find all usages). Furthermore, this allows extracting the dual value of the constraint that fixes the value, assisting in approaches like Benders decomposition. Note that this does not change the induced cost in any way.

fixed_cost	no	-	monetary	-	This setting activates a "fixed cost" component for this decision variable, which requires that the model's problem type allows for binary variables (e.g., MILP). This can be used to model fixed costs that are only incurred if the decision variable is active (e.g., a fixed cost for an investment that is only incurred if the investment is made). If the decision is 0, no fixed costs have to be paid; however, if the decision is greater than 0, the fixed cost is incurred. Note that after deciding to activate the decision, the overall value is still determined in the usual (continuous) way, incuring the (variable) cost as well. More complex cost functions can be modelled by switching to mode sos1 or sos2 and using the sos parameter.

mode	no	linear, binary, integer, sos1, sos2, fixed	-	linear	Type of the decision variable that is constructed. linear results in a continuous decision, integer results in a integer variable, binary constrains it to be either 0 or 1. sos1 and sos2 can be used to activate SOS1 or SOS2 mode (used for piecewise linear costs). See fixed_value if setting this to fixed.

sos	no	list	-	-	TODO (meanwhile, refer to the SOS or PiecewiseLinearCost example).

Detailed Model Reference

Variables

fixed

How to?

Access this variable by using:

# Julia
component(model, "your_decision").var.fixed

# Python
model.get_component("your_decision").var.fixed

You can find the full implementation and all details here: IESopt.jl.

to be added

sos

How to?

Access this variable by using:

# Julia
component(model, "your_decision").var.sos

# Python
model.get_component("your_decision").var.sos

You can find the full implementation and all details here: IESopt.jl.

to be added

value

How to?

Access this variable by using:

# Julia
component(model, "your_decision").var.value

# Python
model.get_component("your_decision").var.value

You can find the full implementation and all details here: IESopt.jl.

Add the variable describing the value of this decision to the model. If lower and upper bounds (decision.lb and decision.ub) are the same, the variable will immediately be fixed to that value. This can be accessed via decision.var.value.

Expressions

Constraints

fixed

How to?

Access this constraint by using:

# Julia
component(model, "your_decision").con.fixed

# Python
model.get_component("your_decision").con.fixed

You can find the full implementation and all details here: IESopt.jl.

to be added

sos1

How to?

Access this constraint by using:

# Julia
component(model, "your_decision").con.sos1

# Python
model.get_component("your_decision").con.sos1

You can find the full implementation and all details here: IESopt.jl.

to be added

sos2

How to?

Access this constraint by using:

# Julia
component(model, "your_decision").con.sos2

# Python
model.get_component("your_decision").con.sos2

You can find the full implementation and all details here: IESopt.jl.

to be added

sos_value

How to?

Access this constraint by using:

# Julia
component(model, "your_decision").con.sos_value

# Python
model.get_component("your_decision").con.sos_value

You can find the full implementation and all details here: IESopt.jl.

to be added

Objectives

fixed

How to?

Access this objective by using:

# Julia
component(model, "your_decision").obj.fixed

# Python
model.get_component("your_decision").obj.fixed

You can find the full implementation and all details here: IESopt.jl.

sos

How to?

Access this objective by using:

# Julia
component(model, "your_decision").obj.sos

# Python
model.get_component("your_decision").obj.sos

You can find the full implementation and all details here: IESopt.jl.

Add the cost defined by the SOS-based value of this Decision to the model.

value

How to?

Access this objective by using:

# Julia
component(model, "your_decision").obj.value

# Python
model.get_component("your_decision").obj.value

You can find the full implementation and all details here: IESopt.jl.

Add the cost defined by the value of this Decision to the model:

\[\text{value} \cdot \text{cost}\]

math

IESopt.Node — Type

A Node represents a basic intersection/hub for energy flows. This can for example be some sort of bus (for electrical systems). It enforces a nodal balance equation (= "energy that flows into it must flow out") for every Snapshot. Enabling the internal state of the Node allows it to act as energy storage, modifying the nodal balance equation. This allows using Nodes for various storage tasks (like batteries, hydro reservoirs, heat storages, ...).

Basic Examples

A Node that represents an electrical bus:

bus:
  type: Node
  carrier: electricity

A Node that represents a simplified hydrogen storage:

store:
  type: Node
  carrier: hydrogen
  has_state: true
  state_lb: 0
  state_ub: 50

Parameters

Name	Mandatory	Values	Unit	Default	Description

carrier	yes	string	-	-	Carrier of this Node. All connecting components need to respect that.

has_state	no	true, false	-	false	If true, the Node is considered to have an internal state ("stateful Node"). This allows it to act as energy storage. Connect Connections or Units to it, acting as charger/discharger.

state_lb	no	numeric, col@file, decision:value	energy	$-\infty$	Lower bound of the internal state, requires has_state = true.

state_ub	no	numeric, col@file, decision:value	energy	$+\infty$	Upper bound of the internal state, requires has_state = true.

state_cyclic	no	eq, geq, or disabled	-	eq	Controls how the state considers the boundary between last and first Snapshot. disabled disables cyclic behaviour of the state (see also state_initial), eq leads to the state at the end of the year being the initial state at the beginning of the year, while geq does the same while allowing the end-of-year state to be higher (= "allowing to destroy energy at the end of the year").

state_initial	no	numeric	energy	-	Sets the initial state. Must be used in combination with state_cyclic = disabled.

state_final	no	numeric	energy	-	Sets the final state. Must be used in combination with state_cyclic = disabled.

state_percentage_loss	no	$\in [0, 1]$	-	0	Per Snapshot percentage loss of state (loosing 1% should be set as 0.01).

nodal_balance	no	enforce, destroy, or create	-	enforce	Can only be used for has_state = false. enforce forces total injections to always be zero (similar to Kirchhoff's current law), create allows "supply < demand", destroy allows "supply > demand", at this Node.

sum_window_size	no	integer	-	-	TODO.

sum_window_step	no	integer	-	1	TODO.

Detailed Model Reference

Variables

pf_theta

How to?

Access this variable by using:

# Julia
component(model, "your_node").var.pf_theta

# Python
model.get_component("your_node").var.pf_theta

You can find the full implementation and all details here: IESopt.jl.

state

How to?

Access this variable by using:

# Julia
component(model, "your_node").var.state

# Python
model.get_component("your_node").var.state

You can find the full implementation and all details here: IESopt.jl.

Add the variable representing the state of this node to the model, if node.has_state == true. This can be accessed via node.var.state[t].

Additionally, if the state's initial value is specified via state_initial the following gets added:

\[\text{state}_1 = \text{state}_{initial}\]

math

Expressions

injection

How to?

Access this expression by using:

# Julia
component(model, "your_node").exp.injection

# Python
model.get_component("your_node").exp.injection

You can find the full implementation and all details here: IESopt.jl.

Add an empty (JuMP.AffExpr(0)) expression to the node that keeps track of feed-in and withdrawal of energy.

This constructs the expression $\text{injection}_t, \forall t \in T$ that is utilized in node.con.nodalbalance. Core components (Connections, Profiles, and Units) that feed energy into this node add to it, all others subtract from it. A stateless node forces this nodal balance to always equal 0 which essentially describes "generation = demand".

Constraints

last_state

How to?

Access this constraint by using:

# Julia
component(model, "your_node").con.last_state

# Python
model.get_component("your_node").con.last_state

You can find the full implementation and all details here: IESopt.jl.

Add the constraint defining the bounds of the node's state during the last Snapshot to the model, if node.has_state == true.

This is necessary since it could otherwise happen, that the state following the last Snapshot is actually not feasible (e.g. we could charge a storage by more than it's state allows for). The equations are based on the construction of the overall state variable.

\[\begin{aligned} & \text{state}_{end} \cdot \text{factor}^\omega_t + \text{injection}_{end} \cdot \omega_t \geq \text{state}_{lb} \\ & \text{state}_{end} \cdot \text{factor}^\omega_t + \text{injection}_{end} \cdot \omega_t \leq \text{state}_{ub} \end{aligned}\]

math

Here $\omega_t$ is the weight of Snapshot t, and $\text{factor}$ is either 1.0 (if there are now percentage losses configured), or (1.0 - node.state_percentage_loss) otherwise.

Constraint safety

The lower and upper bound constraint are subject to penalized slacks.

nodalbalance

How to?

Access this constraint by using:

# Julia
component(model, "your_node").con.nodalbalance

# Python
model.get_component("your_node").con.nodalbalance

You can find the full implementation and all details here: IESopt.jl.

Add the constraint describing the nodal balance to the model.

Depending on whether the node is stateful or not, this constructs different representations:

if node.has_state == true

\[\begin{aligned} & \text{state}_t = \text{state}_{t-1} \cdot \text{factor}^\omega_{t-1} + \text{injection}_{t-1} \cdot \omega_{t-1}, \qquad \forall t \in T \setminus \{1\} \\ \\ & \text{state}_1 = \text{state}_{end} \cdot \text{factor}^\omega_{end} + \text{injection}_{end} \cdot \omega_{end} \end{aligned}\]

math

Here $\omega_t$ is the weight of Snapshot t, and $\text{factor}$ is either 1.0 (if there are now percentage losses configured), or (1.0 - node.state_percentage_loss) otherwise. $\text{injection}_{t}$ describes the overall injection (all feed-ins minus all withdrawals). $end$ indicates the last snapshot in $T$. Depending on the setting of state_cyclic the second constraint is written as $=$ ("eq") or $\leq$ ("leq"). The latter allows the destruction of excess energy at the end of the total time period to help with feasibility.

if node.has_state == false

\[\begin{aligned} & \text{injection}_{t} = 0, \qquad \forall t \in T \\ \end{aligned}\]

math

This equation can further be configured using the nodal_balance parameter, which accepts enforce (resulting in $=$), create (resulting in $\leq$; allowing the creation of energy - or "negative injections"), and destroy ( resulting in $\geq$; allowing the destruction of energy - or "positive injections"). This can be used to model some form of energy that can either be sold (using a destroy Profile connected to this Node), or "wasted into the air" using the destroy setting of this Node.

state_bounds

How to?

Access this constraint by using:

# Julia
component(model, "your_node").con.state_bounds

# Python
model.get_component("your_node").con.state_bounds

You can find the full implementation and all details here: IESopt.jl.

Add the constraint defining the bounds of the node's state to the model, if node.has_state == true.

\[\begin{aligned} & \text{state}_t \geq \text{state}_{lb}, \qquad \forall t \in T \\ & \text{state}_t \leq \text{state}_{ub}, \qquad \forall t \in T \end{aligned}\]

math

Constraint safety

The lower and upper bound constraint are subject to penalized slacks.

Objectives

IESopt.Profile — Type

A Profile allows representing "model boundaries" - parts of initial problem that are not endogenously modelled - with a support for time series data. Examples are hydro reservoir inflows, electricity demand, importing gas, and so on. Besides modelling fixed profiles, they also allow different ways to modify the value endogenously.

Basic Examples

A Profile that depicts a fixed electricity demand:

demand_XY:
  type: Profile
  carrier: electricity
  node_from: grid
  value: demand_XY@input_file

A Profile that handles cost of fuel:

fuel_gas:
  type: Profile
  carrier: gas
  node_to: country_gas_grid
  mode: create
  cost: 100.0

A Profile that handles CO2 emission costs:

co2_cost:
  type: Profile
  carrier: co2
  node_from: total_co2
  mode: destroy
  cost: 150.0

A Profile that handles selling electricity:

sell_electricity:
  type: Profile
  carrier: electricity
  node_from: internal_grid_node
  mode: destroy
  cost: -30.0

Parameters

Name	Mandatory	Values	Unit	Default	Description

carrier	yes	string	-	-	Carrier of this Profile. Must match the Carrier of the Node that this connects to.

value	no	numeric, col@file	power	-	The concrete value of this Profile - either static or as time series. Only applicable if mode: fixed.

node_from	no	string	-	-	Name of the Node that this Profile draws energy from. Exactly one of node_from and node_to must be set.

node_to	no	string	-	-	Name of the Node that this Profile feeds energy to. Exactly one of node_from and node_to must be set.

mode	no	-	-	fixed	The mode of operation of this Profile. fixed uses the supplied value, ranged allows ranging between lb and ub, while create (must specify node_to) and destroy (must specify node_from) handle arbitrary energy flows that are bounded from below by 0. Use fixed if you want to fix the value of the Profile to a specific value, e.g., a given energy demand. Use create to "import" energy into the model, e.g., from a not explicitly modelled gas market, indcucing a certain cost for buying that energy. Use destroy to "export" energy from the model, e.g., to handle CO2 going into the atmosphere (which may be taxed, etc., by the cost of this Profile). Use ranged if you need more fine grained control over the value of the Profile, than what create and destroy allow (e.g., a grid limited energy supplier).

lb	no	numeric	power	$-\infty$	The lower bound of the range of this Profile (must be used together with mode: ranged).

ub	no	numeric	power	$+\infty$	The upper bound of the range of this Profile (must be used together with mode: ranged).

cost	no	numeric	monetary per energy	0	Cost per unit of energy that this Profile injects or withdraws from a Node. Refer to the basic examples to see how this can be combined with mode for different use cases.

Detailed Model Reference

Variables

aux_value

How to?

Access this variable by using:

# Julia
component(model, "your_profile").var.aux_value

# Python
model.get_component("your_profile").var.aux_value

You can find the full implementation and all details here: IESopt.jl.

Add the variable that is used in this Profiles value to the model.

The variable var_value[t] is constructed and is linked to the correct Nodes. There are different ways, IESopt interprets this, based on the setting of profile.mode:

fixed: The value is already handled by the constant term of profile.exp.value and NO variable is constructed.

create, destroy, or ranged: This models the creation or destruction of energy - used mainly to represent model boundaries, and energy that comes into the model or leaves the model's scope. It is however important that create should mostly be used feeding into a Node (profile.node_from = nothing) and destroy withdrawing from a Node (profile.node_to = nothing). If lb and ub are defined, ranged can be used that allows a more detailled control over the Profile, specifying upper and lower bounds for every Snapshot. See _profile_con_value_bounds!(profile::Profile) for details on the specific bounds for each case.

This variable is added to the profile.exp.value. Additionally, the energy (that profile.exp.value represents) gets "injected" at the Nodes that the profile is connected to, resulting in

\[\begin{aligned} & \text{profile.node}_{from}\text{.injection}_t = \text{profile.node}_{from}\text{.injection}_t - \text{value}_t, \qquad \forall t \in T \\ & \text{profile.node}_{to}\text{.injection}_t = \text{profile.node}_{to}\text{.injection}_t + \text{value}_t, \qquad \forall t \in T \end{aligned}\]

math

Expressions

value

How to?

Access this expression by using:

# Julia
component(model, "your_profile").exp.value

# Python
model.get_component("your_profile").exp.value

You can find the full implementation and all details here: IESopt.jl.

Cosntruct the JuMP.AffExpr that keeps the total value of this Profile for each Snapshot.

This is skipped if the value of this Profile is handled by an Expression. Otherwise it is intialized based on profile.value.

Constraints

value_bounds

How to?

Access this constraint by using:

# Julia
component(model, "your_profile").con.value_bounds

# Python
model.get_component("your_profile").con.value_bounds

You can find the full implementation and all details here: IESopt.jl.

Add the constraint defining the bounds of this profile to the model.

This heavily depends on the mode setting, as it does nothing if the mode is set to fixed, or the value is actually controlled by an Expression. The variable can be accessed via profile.var.aux_value[t], but using the normal result extraction is recommended, since that properly handles the profile.exp.value instead.

Otherwise:

if profile.mode === :create or profile.mode === :destroy

\[\begin{aligned} & \text{aux_value}_t \geq 0, \qquad \forall t \in T \end{aligned}\]

math

if profile.mode === :ranged

\[\begin{aligned} & \text{value}_t \geq \text{lb}_t, \qquad \forall t \in T \\ & \text{value}_t \leq \text{ub}_t, \qquad \forall t \in T \end{aligned}\]

math

Here, lb and ub can be left empty, which drops the respective constraint.

Objectives

cost

How to?

Access this objective by using:

# Julia
component(model, "your_profile").obj.cost

# Python
model.get_component("your_profile").obj.cost

You can find the full implementation and all details here: IESopt.jl.

Add the (potential) cost of this Profile to the global objective function.

The profile.cost setting specifies a potential cost for the creation ("resource costs", i.e. importing gas into the model) or destruction ("penalties", i.e. costs linked to the emission of CO2). It can have a unique value for every Snapshot, i.e. allowing to model a time-varying gas price throughout the year.

The contribution to the global objective function is as follows:

\[\sum_{t\in T} \text{value}_t \cdot \text{profile.cost}_t \cdot \omega_t\]

math

Here $\omega_t$ is the weight of Snapshot t, and $\text{value}_t$ actually refers to the value of profile.exp.value[t] (and not only on the maybe non-existing variable).

IESopt.Unit — Type

A Unit allows transforming one (or many) forms of energy into another one (or many), given some constraints and costs.

Basic Examples

A Unit that represents a basic gas turbine:

gas_turbine:
  type: Unit
  inputs: {gas: gas_grid}
  outputs: {electricity: node, co2: total_co2}
  conversion: 1 gas -> 0.4 electricity + 0.2 co2
  capacity: 10 out:electricity

A Unit that represents a basic wind turbine:

wind_turbine:
  type: Unit
  outputs: {electricity: node}
  conversion: ~ -> 1 electricity
  capacity: 10 out:electricity
  availability_factor: wind_factor@input_data
  marginal_cost: 1.7 per out:electricity

A Unit that represents a basic heat pump, utilizing a varying COP:

heatpump:
  type: Unit
  inputs: {electricity: grid}
  outputs: {heat: heat_system}
  conversion: 1 electricity -> cop@inputfile heat
  capacity: 10 in:electricity

Parameters

Name	Mandatory	Values	Unit	Default	Description

conversion	yes	string	-	-	The conversion expression describing how this Unit transforms energy. Specified in the form of "$\alpha \cdot carrier_1 + \beta \cdot carrier_2$ -> $\gamma \cdot carrier_3 + \delta \cdot carrier_4$". Coefficients allow simple numerical calculations, but are not allowed to include spaces (so e.g. (1.0/9.0) is valid). Coefficients are allowed to be NumericalInputs, resulting in column@data_file being a valid coefficient (this can be used e.g. for time-varying COPs of heatpumps).

capacity	yes	value dir:carrier	-	-	Maximum capacity of this Unit, to be given in the format X in/out:carrier where X is the amount, in or out (followed by :) specifies whether the limit is to be placed on the in- our output of this Unit, and carrier specifies the respective Carrier. Example: 100 in:electricity (to limit the "input rating").

outputs	yes	dict	-	-	Dictionary specifying the output "ports" of this Unit. Refer to the basic examples for the general syntax.

inputs	no	dict	-	-	Dictionary specifying the input "ports" of this Unit. If not specified (= no explicit input), the conversion has to follow the form of conversion: ~ -> ..., indicating an "open" input. This may, e.g., be used for renewable energy sources, where the primary energy input (e.g., solar) is not explicitly modeled.

availability	no	numeric	power	$+\infty$	Time series (or fixed value) that limits the available capacity. If, e.g., capacity: 100 out:electricity and availability: 70, the available capacity will only be 50 electricity. Can be used to model non-availability of power plants, e.g., due to maintenance. For time-varying availability of intermittent generators (e.g., wind), it's recommended (most of the time) to use availability_factor instead.

availability_factor	no	$\in [0, 1]$	-	1	Similar to availability, but given as factor of capacity instead. If, e.g., capacity: 100 out:electricity and availability_factor: 0.7, the available capacity will only be 70 electricity. This is especially useful for intermittent generators, where the availability is not a fixed value, but depends on the weather, and can be passed, e.g., by setting availability_factor: wind@input_data_file.

adapt_min_to_availability	no	true, false	-	false	If true, the minimal partial load will be influenced by the availability. Example: Consider a Unit with capacity: 100 out:electricity, a min_conversion of 0.4, and an availability_factor of 0.5. This entails having 50 electricity available, while the minimal partial load is 40 electricity. This results in the Unit at best operating only closely above the minimal partial load. Furthermore, an availability_factor below 0.4 would result in no feasible generation, besides shutting the Unit off. While this might be the intended mode of operation in many use cases, adapt_min_to_availability can change this: If set to true, this dynamically changes the minimal partial load. In the previous example, that means (100 * 0.5) * 0.4 = 20 electricity (the 50% minimum load are now based on the available 40), changing the overall behaviour (including efficiencies) as well as leading to feasible generations even when the availability_factor is below 0.4.

marginal_cost	no	value per dir:carrier	monetary per energy	0	Marginal cost of the consumption/generation of one unit of energy of the specified carrier. Has to be given in the format value per dir:carrier, e.g. 3.5 per out:electricity for a marginal cost of 3.5 monetary units per unit of electricity generated.

enable_ramp_up	no	true, false	-	false	Enables calculation of upward ramps. Ramping is based on the carrier specified in capacity.

enable_ramp_down	no	true, false	-	false	Enables calculation of downward ramps. Ramping is based on the carrier specified in capacity.

ramp_up_cost	no	numeric	monetary per power	0	Sets the cost of ramping up (increasing in-/output) by 1 unit of the capacity carrier.

ramp_down_cost	no	numeric	monetary per power	0	Sets the cost of ramping down (decreasing in-/output) by 1 unit of the capacity carrier.

ramp_up_limit	no	$\in [0, 1]$	-	1	Limits the allowed ramping up based on this factor of the total capacity. If capacity: 100 in:electricity with ramp_up_limit: 0.2, this limits the total increase of usage of electricity (on the input) to 20 units (power) per hour. For example, starting at an input of 35, after one hour the input has to be lesser than or equal to 55. If a Snapshot's duration is set to, e.g., two hours, this would allow a total increase of 40 units.

ramp_down_limit	no	$\in [0, 1]$	-	1	Limits the allowed ramping down based on this factor of the total capacity. See ramp_up_limit.

min_on_time	no	numeric	hours	0	Minimum on-time of the Unit. If set, the Unit has to be on for at least this amount of time, after turning on. It is highly recommended to only use this with unit_commitment: binary, unless you know why it's fine to use with another mode.

min_off_time	no	numeric	hours	0	Minimum off-time of the Unit. If set, the Unit has to be off for at least this amount of time, after turning off. It is highly recommended to only use this with unit_commitment: binary, unless you know why it's fine to use with another mode.

on_time_before	no	numeric	hours	0	Time that this Unit has already been running before the optimization starts. Can be used in combination with min_on_time.

off_time_before	no	numeric	hours	0	Time that this Unit has already been off before the optimization starts. Can be used in combination with min_off_time.

is_on_before	no	numeric	-	1	Number of Units that should be considered to have been running before the optimization starts. Can be used in combination with on_time_before, especially for unit_count greater than 1.

unit_commitment	no	off, linear, binary, integer	-	off	Controls how the unit commitment of this Unit is handled. linear results in the ability to startup parts of the unit (so 0.314159 is a feasible amount of "turned on unit"), while binary restricts the Unit to either be on (converting the conversion_at_min + possible additional conversion above that minimum) or off (converting nothing); integer is needed to consider binary unit commitment for Units with more than 1 "grouped unit" (see unit_count).

unit_count	no	numeric	-	1	Number of units aggregated in this Unit. Besides interacting with the mode of unit_commitment, this mainly is responsible for scaling the output (e.g. grouping 47 of the same wind turbine, ...).

min_conversion	no	$\in [0, 1]$	-	-	If unit_commitment is not set to off, this specifies the percentage that is considered to be the minimal feasible partial load this Unit can operate at. Operating below that setpoint is not allowed, at that point the conversion_at_min coefficients are used, and above that they are scaled to result in conversion when running at full capacity.

conversion_at_min	no	string	-	-	The conversion expression while running on the minimal partial load. Only applicable if unit_commitment is not off and min_conversion is explicitly set. Follows the same form as conversion.

startup_cost	no	numeric	monetary per start	0	Costs per startup (also applicable if startups are not binary or integer). This is necessary to allow conversion_at_min to have (at least partially) the effect that one expects, if unit_commitment: linear.

Detailed Model Reference

Variables

conversion

How to?

Access this variable by using:

# Julia
component(model, "your_unit").var.conversion

# Python
model.get_component("your_unit").var.conversion

You can find the full implementation and all details here: IESopt.jl.

Add the variable describing the unit's conversion to the model.

This can be accessed via unit.var.conversion[t]; this does not describe the full output of the Unit since that maybe also include fixed generation based on the ison variable.

Info

This applies some heavy recalculation of efficiencies to account for minimum load and so on, that are currently not fully documented. This essentially comes down to the following: As long as minimum load is not enabled, that is rather simple (using the conversion expression to withdraw energy from the inputs and push energy into the outputs). If a separate minimum load conversion is specified it results in the following: (1) if running at minimum load the supplied minimum load conversion will be used; (2) if running at maximum capacity the "normal" conversion expression will be used; (3) for any point in-between a linear interpolation scales up all coefficients of the conversion expression to "connect" case (1) and (2).

conversion_connect

No documentation found.

IESopt._unit_var_conversion_connect! is a Function.

How to?

Access this variable by using:

# Julia
component(model, "your_unit").var.conversion_connect

# Python
model.get_component("your_unit").var.conversion_connect

You can find the full implementation and all details here: IESopt.jl.

ison

How to?

Access this variable by using:

# Julia
component(model, "your_unit").var.ison

# Python
model.get_component("your_unit").var.ison

You can find the full implementation and all details here: IESopt.jl.

Add the variable describing the current "online" state of the unit to the model.

The variable can be further parameterized using the unit.unit_commitment setting ("linear", "binary", "integer"). It will automatically enforce the constraints $0 \leq \text{ison} \leq \text{unitcount}$, with $\text{unitcount}$ describing the number of units that are aggregated in this unit (set by unit.unit_count). This can be accessed via unit.var.ison[t].

ramp

How to?

Access this variable by using:

# Julia
component(model, "your_unit").var.ramp

# Python
model.get_component("your_unit").var.ramp

You can find the full implementation and all details here: IESopt.jl.

Add the variable describing the per-snapshot ramping to the model.

This adds two variables per snapshot to the model (if the respective setting unit.enable_ramp_up or unit.enable_ramp_down is activated). Both are preconstructed with a fixed lower bound of 0. This describes the amount of change in conversion that occurs during the current snapshot. These can be accessed via unit.var.ramp_up[t] and unit.var.ramp_down[t].

These variables are only used for ramping costs. The limits are enforced directly on the conversion, which means this variable only exists if costs are specified!

startup

How to?

Access this variable by using:

# Julia
component(model, "your_unit").var.startup

# Python
model.get_component("your_unit").var.startup

You can find the full implementation and all details here: IESopt.jl.

Add the variable describing the per-snapshot startup to the model.

This adds a variable per snapshot to the model (if the respective setting unit.unit_commitment is activated). The variable can be further parameterized using the unit.unit_commitment setting ("linear", "binary", "integer"). It will automatically enforce the constraints $0 \leq \text{startup} \leq \text{unitcount}$, with $\text{unitcount}$ describing the number of units that are aggregated in this unit (set by unit.unit_count). This describes the startup that happens during the current snapshot and can be accessed via unit.var.startup.

Expressions

Constraints

conversion_bounds

How to?

Access this constraint by using:

# Julia
component(model, "your_unit").con.conversion_bounds

# Python
model.get_component("your_unit").con.conversion_bounds

You can find the full implementation and all details here: IESopt.jl.

Add the constraint defining the unit's conversion bounds to the model.

This makes use of the current min_capacity (describing the lower limit of conversion; either 0 if no minimum load applies or the respective value of the minimum load) as well as the online_capacity (that can either be the full capacity if unit commitment is disabled, or the amount that is currently active).

Depending on how the "availability" of this unit is handled it constructs the following constraints:

if !isnothing(unit.availability)

\[\begin{aligned} & \text{conversion}_t \geq \text{capacity}_{\text{min}, t}, \qquad \forall t \in T \\ & \text{conversion}_t \leq \text{capacity}_{\text{online}, t}, \qquad \forall t \in T \\ & \text{conversion}_t \leq \text{availability}_t, \qquad \forall t \in T \end{aligned}\]

math

This effectively results in $\text{conversion}_t \leq \min(\text{capacity}_{\text{online}, t}, \text{availability}_t)$.

if !isnothing(unit.availability_factor)

\[\begin{aligned} & \text{conversion}_t \geq \text{capacity}_{\text{min}, t}, \qquad \forall t \in T \\ & \text{conversion}_t \leq \text{capacity}_{\text{online}, t} \cdot \text{availability}_{\text{factor}, t}, \qquad \forall t \in T \end{aligned}\]

math

Info

If one is able to choose between using availability or availability_factor (e.g. for restricting available capacity during a planned revision to half the units capacity), enabling availability_factor (in this example 0.5) will result in a faster model (build and probably solve) since it makes use of one less constraint.

If no kind of availability limiting takes place, the following bounds are enforced:

\[\begin{aligned} & \text{conversion}_t \geq \text{capacity}_{\text{min}, t}, \qquad \forall t \in T \\ & \text{conversion}_t \leq \text{capacity}_{\text{online}, t}, \qquad \forall t \in T \end{aligned}\]

math

ison

How to?

Access this constraint by using:

# Julia
component(model, "your_unit").con.ison

# Python
model.get_component("your_unit").con.ison

You can find the full implementation and all details here: IESopt.jl.

Construct the upper bound for var_ison, based on unit.unit_count, if it is handled by an external Decision.

min_onoff_time

How to?

Access this constraint by using:

# Julia
component(model, "your_unit").con.min_onoff_time

# Python
model.get_component("your_unit").con.min_onoff_time

You can find the full implementation and all details here: IESopt.jl.

Add the constraints modeling min on- or off-time of a Unit to the model.

This constructs the constraints

\[\begin{align} & \sum_{t' = t}^{t + \text{min\_on\_time}} ison_{t'} >= \text{min\_on\_time} \cdot (ison_t - ison_{t-1}) \qquad \forall t \in T \\ & \sum_{t' = t}^{t + \text{min\_off\_time}} (1 - ison_{t'}) >= \text{min\_off\_time} \cdot (ison_{t-1} - ison_t) \qquad \forall t \in T \end{align} respecting `on_time_before` and `off_time_before`, and `is_on_before`. See the code for more details. !!! info "Aggregated units" This is currently not fully adapted to account for `Unit`s with `unit_count > 1`.\]

math

ramp

How to?

Access this constraint by using:

# Julia
component(model, "your_unit").con.ramp

# Python
model.get_component("your_unit").con.ramp

You can find the full implementation and all details here: IESopt.jl.

Add the auxiliary constraint that enables calculation of per snapshot ramping to the model.

Depending on whether ramps are enabled, none, one, or both of the following constraints are constructed:

\[\begin{aligned} & \text{ramp}_{\text{up}, t} \geq \text{conversion}_{t} - \text{conversion}_{t-1}, \qquad \forall t \in T \\ & \text{ramp}_{\text{down}, t} \geq \text{conversion}_{t-1} - \text{conversion}_{t}, \qquad \forall t \in T \end{aligned}\]

math

This calculates the ramping that happens from the PREVIOUS snapshot to this one. That means that if:

out[5] = 100 and out[4] = 50, then ramp_up[5] = 50 and ramp_down[5] = 0

ramp_up[1] = ramp_down[1] = 0

Info

This currently does not support pre-setting the initial states of the unit (it can be done manually but there is no exposed parameter), which will be implemented in the future to allow for easy / correct rolling optimization runs.

ramp_limit

How to?

Access this constraint by using:

# Julia
component(model, "your_unit").con.ramp_limit

# Python
model.get_component("your_unit").con.ramp_limit

You can find the full implementation and all details here: IESopt.jl.

Add the constraint describing the ramping limits of this unit to the model.

This makes use of the maximum capacity of the unit, which is just the total installed capacity. Both, up- and downwards ramps can be enabled separately (via unit.ramp_up_limit and unit.ramp_down_limit), resulting in either or both of:

\[\begin{aligned} & \text{ramp}_{\text{up}, t} \leq \text{ramplimit}_\text{up} \cdot \text{capacity}_\text{max} \cdot \omega_t, \qquad \forall t \in T \\ & \text{ramp}_{\text{down}, t} \leq \text{ramplimit}_\text{down} \cdot \text{capacity}_\text{max} \cdot \omega_t, \qquad \forall t \in T \end{aligned}\]

math

This does not make use of the ramping variable (that is only used for costs - if there are costs).

This calculates the ramping that happens from the PREVIOUS snapshot to this one. That means that if:

out[5] = 100 and out[4] = 50, then ramp_up[5] = 50 and ramp_down[5] = 0

ramp_up[1] = ramp_down[1] = 0

startup

How to?

Access this constraint by using:

# Julia
component(model, "your_unit").con.startup

# Python
model.get_component("your_unit").con.startup

You can find the full implementation and all details here: IESopt.jl.

Add the auxiliary constraint that enables calculation of per snapshot startup to the model.

Depending on whether startup handling is enabled, the following constraint is constructed:

\[\begin{aligned} & \text{startup}_{\text{up}, t} \geq \text{ison}_{t} - \text{ison}_{t-1}, \qquad \forall t \in T \end{aligned}\]

math

This calculates the startup that happens from the PREVIOUS snapshot to this one. That means that if:

ison[5] = 1 and ison[4] = 0, then startup[5] = 1

Objectives

marginal_cost

How to?

Access this objective by using:

# Julia
component(model, "your_unit").obj.marginal_cost

# Python
model.get_component("your_unit").obj.marginal_cost

You can find the full implementation and all details here: IESopt.jl.

Add the (potential) cost of this unit's conversion (unit.marginal_cost) to the global objective function.

\[\sum_{t \in T} \text{conversion}_t \cdot \text{marginalcost}_t \cdot \omega_t\]

math

ramp_cost

How to?

Access this objective by using:

# Julia
component(model, "your_unit").obj.ramp_cost

# Python
model.get_component("your_unit").obj.ramp_cost

You can find the full implementation and all details here: IESopt.jl.

Add the (potential) cost of this unit's ramping to the global objective function.

To allow for finer control, costs of up- and downwards ramping can be specified separately (using unit.ramp_up_cost and unit.ramp_down_cost):

\[\sum_{t \in T} \text{ramp}_{\text{up}, t} \cdot \text{rampcost}_{\text{up}} + \text{ramp}_{\text{down}, t} \cdot \text{rampcost}_{\text{down}}\]

math

startup_cost

How to?

Access this objective by using:

# Julia
component(model, "your_unit").obj.startup_cost

# Python
model.get_component("your_unit").obj.startup_cost

You can find the full implementation and all details here: IESopt.jl.

Add the (potential) cost of this unit's startup behaviour (configured by unit.startup_cost if unit.unit_commitment != :off).

\[\sum_{t \in T} \text{startup}_t \cdot \text{startupcost}\]

math