Model formulation

This page collects all information about the core components of IESopt, their properties (and default values), their internal mathematical formulations, and some examples of how to use them. See the API for a complete list of all available functions and types.

IESopt.Connection — Type

nothing

Parameters

Name	Mandatory	Values	Unit	Default	Description

node_from	yes	string	-	-	This Connection models a flow from node_from to node_to (both are Nodes).

node_to	yes	string	-	-	This Connection models a flow from node_from to node_to (both are Nodes).

carrier	no	string	-	-	Carrier of this Connection. If not given, automatically picks the carrier of the Nodes it connects. This parameter is not necessary, and only exists to allow for a more explicit definition.

capacity	no	numeric, col@file, decision:value	power	$+\infty$	The symmetric bound on this Connection's flow. Results in lb = -capacity and ub = capacity. Must not be specified if lb, ub, or both are explicitly stated.

lb	no	numeric, col@file, decision:value	power	$-\infty$	Lower bound of this Connection's flow.

ub	no	numeric, col@file, decision:value	power	$+\infty$	Upper bound of this Connection's flow.

cost	no	numeric	monetary (per energy)	-	Cost of every unit of energy flow over this connection that is added to the model's objective function. Keep in mind that negative flows will induce negative costs, which can be used to model revenues. Further, a bidirectional Connection (if lb < 0, which is the default, or if capacity is used) with a positive cost will lead to negative costs for the reverse flow. If you do not want this, split the Connection into two separate ones, each being unidirectional (with lb: 0). Remember, that these can share the same "capacity" (which is then set asub), even when using decision:value or col@file as value.

loss	no	$\in [0, 1]$	-	0	Fractional loss when transferring energy. This loss occurs "at the destination", which means that for a loss of 5%, set as loss: 0.05, and considering a Snapshot where the Connection has a flow value of 100, it will "extract" 100 from node_from and "inject" 95 into node_to. Since the flow variable is given as power, this would, e.g., translate to consuming 200 units of energy at node_from and injecting 190 units at node_to, if the Snapshot duration is 2 hours.

build_priority	no	numeric	-	0	Priority for the build order of components. Components with higher build_priority are built before. This can be useful for addons, that connect multiple components and rely on specific components being initialized before others.

Detailed Model Reference

Variables

flow

nothing

Expressions

pf_flow

nothing

Constraints

flow_bounds

nothing

Objectives

cost

nothing

IESopt.Decision — Type

nothing

Parameters

Name	Mandatory	Values	Unit	Default	Description

lb	no	numeric	-	0	Minimum size of the decision value (considered for each "unit" if count allows multiple "units").

ub	no	numeric	-	$+\infty$	Maximum size of the decision value (considered for each "unit" if count allows multiple "units").

cost	no	numeric	monetary (per value)	0	Cost that the decision value induces, given as $cost \cdot value$.

fixed_value	no	numeric	-	-	If mode: fixed, this value is used as the fixed value of the decision. This can be useful if this Decision was used in a previous optimization and its value should be fixed to that value in the next optimization (applying it where ever it is used, instead of needing to find all usages). Furthermore, this allows extracting the dual value of the constraint that fixes the value, assisting in approaches like Benders decomposition. Note that this does not change the induced cost in any way.

fixed_cost	no	-	monetary	-	This setting activates a "fixed cost" component for this decision variable, which requires that the model's problem type allows for binary variables (e.g., MILP). This can be used to model fixed costs that are only incurred if the decision variable is active (e.g., a fixed cost for an investment that is only incurred if the investment is made). If the decision is 0, no fixed costs have to be paid; however, if the decision is greater than 0, the fixed cost is incurred. Note that after deciding to activate the decision, the overall value is still determined in the usual (continuous) way, incuring the (variable) cost as well. More complex cost functions can be modelled by switching to mode sos1 or sos2 and using the sos parameter.

mode	no	linear, binary, integer, sos1, sos2, fixed	-	linear	Type of the decision variable that is constructed. linear results in a continuous decision, integer results in a integer variable, binary constrains it to be either 0 or 1. sos1 and sos2 can be used to activate SOS1 or SOS2 mode (used for piecewise linear costs). See fixed_value if setting this to fixed.

sos	no	list	-	-	TODO (meanwhile, refer to the SOS or PiecewiseLinearCost example).

build_priority	no	numeric	-	1000	Priority for the build order of components. Components with higher build_priority are built before. This can be useful for addons, that connect multiple components and rely on specific components being initialized before others.

Detailed Model Reference

Variables

fixed

nothing

sos

nothing

value

nothing

Expressions

Constraints

fixed

nothing

sos1

nothing

sos2

nothing

sos_value

nothing

Objectives

fixed

nothing

sos

nothing

value

nothing

IESopt.Node — Type

nothing

Parameters

Name	Mandatory	Values	Unit	Default	Description

carrier	yes	string	-	-	Carrier of this Node. All connecting components need to respect that.

has_state	no	true, false	-	false	If true, the Node is considered to have an internal state ("stateful Node"). This allows it to act as energy storage. Connect Connections or Units to it, acting as charger/discharger.

state_lb	no	numeric, col@file, decision:value	energy	$-\infty$	Lower bound of the internal state, requires has_state = true.

state_ub	no	numeric, col@file, decision:value	energy	$+\infty$	Upper bound of the internal state, requires has_state = true.

state_cyclic	no	eq, geq, or disabled	-	eq	Controls how the state considers the boundary between last and first Snapshot. disabled disables cyclic behaviour of the state (see also state_initial), eq leads to the state at the end of the year being the initial state at the beginning of the year, while geq does the same while allowing the end-of-year state to be higher (= "allowing to destroy energy at the end of the year").

state_initial	no	numeric	energy	-	Sets the initial state. Must be used in combination with state_cyclic = disabled.

state_final	no	numeric	energy	-	Sets the final state. Must be used in combination with state_cyclic = disabled.

state_percentage_loss	no	$\in [0, 1]$	-	0	Per Snapshot percentage loss of state (losing 1% should be set as 0.01).

nodal_balance	no	enforce, destroy, or create	-	enforce	Can only be used for has_state = false. enforce forces total injections to always be zero (similar to Kirchhoff's current law), create allows "supply < demand", destroy allows "supply > demand", at this Node.

sum_window_size	no	integer	-	-	TODO.

sum_window_step	no	integer	-	1	TODO.

build_priority	no	numeric	-	0	Priority for the build order of components. Components with higher build_priority are built before. This can be useful for addons, that connect multiple components and rely on specific components being initialized before others.

Detailed Model Reference

Variables

pf_theta

nothing

state

nothing

Expressions

injection

nothing

Constraints

last_state

nothing

nodalbalance

nothing

state_bounds

nothing

Objectives

IESopt.Profile — Type

nothing

Parameters

Name	Mandatory	Values	Unit	Default	Description

carrier	yes	string	-	-	Carrier of this Profile. Must match the Carrier of the Node that this connects to.

value	no	numeric, col@file	power	-	The concrete value of this Profile - either static or as time series. Only applicable if mode: fixed.

node_from	no	string	-	-	Name of the Node that this Profile draws energy from. Exactly one of node_from and node_to must be set.

node_to	no	string	-	-	Name of the Node that this Profile feeds energy to. Exactly one of node_from and node_to must be set.

mode	no	-	-	fixed	The mode of operation of this Profile. fixed uses the supplied value, ranged allows ranging between lb and ub, while create (must specify node_to) and destroy (must specify node_from) handle arbitrary energy flows that are bounded from below by 0. Use fixed if you want to fix the value of the Profile to a specific value, e.g., a given energy demand. Use create to "import" energy into the model, e.g., from a not explicitly modelled gas market, indcucing a certain cost for buying that energy. Use destroy to "export" energy from the model, e.g., to handle CO2 going into the atmosphere (which may be taxed, etc., by the cost of this Profile). Use ranged if you need more fine grained control over the value of the Profile, than what create and destroy allow (e.g., a grid limited energy supplier).

lb	no	numeric	power	$-\infty$	The lower bound of the range of this Profile (must be used together with mode: ranged).

ub	no	numeric	power	$+\infty$	The upper bound of the range of this Profile (must be used together with mode: ranged).

cost	no	numeric	monetary per energy	0	Cost per unit of energy that this Profile injects or withdraws from a Node. Refer to the basic examples to see how this can be combined with mode for different use cases.

build_priority	no	numeric	-	0	Priority for the build order of components. Components with higher build_priority are built before. This can be useful for addons, that connect multiple components and rely on specific components being initialized before others.

Detailed Model Reference

Variables

aux_value

nothing

Expressions

value

nothing

Constraints

value_bounds

nothing

Objectives

cost

nothing

IESopt.Unit — Type

nothing

Parameters

Name	Mandatory	Values	Unit	Default	Description

conversion	yes	string	-	-	The conversion expression describing how this Unit transforms energy. Specified in the form of "$\alpha \cdot carrier_1 + \beta \cdot carrier_2$ -> $\gamma \cdot carrier_3 + \delta \cdot carrier_4$". Coefficients allow simple numerical calculations, but are not allowed to include spaces (so e.g. (1.0/9.0) is valid). Coefficients are allowed to be NumericalInputs, resulting in column@data_file being a valid coefficient (this can be used e.g. for time-varying COPs of heatpumps).

capacity	yes	value dir:carrier	-	-	Maximum capacity of this Unit, to be given in the format X in/out:carrier where X is the amount, in or out (followed by :) specifies whether the limit is to be placed on the in- our output of this Unit, and carrier specifies the respective Carrier. Example: 100 in:electricity (to limit the "input rating").

outputs	yes	dict	-	-	Dictionary specifying the output "ports" of this Unit. Refer to the basic examples for the general syntax.

inputs	no	dict	-	-	Dictionary specifying the input "ports" of this Unit. If not specified (= no explicit input), the conversion has to follow the form of conversion: ~ -> ..., indicating an "open" input. This may, e.g., be used for renewable energy sources, where the primary energy input (e.g., solar) is not explicitly modeled.

availability	no	numeric	power	$+\infty$	Time series (or fixed value) that limits the available capacity. If, e.g., capacity: 100 out:electricity and availability: 70, the available capacity will only be 70 electricity. Can be used to model non-availability of power plants, e.g., due to maintenance. For time-varying availability of intermittent generators (e.g., wind), it's recommended (most of the time) to use availability_factor instead.

availability_factor	no	$\in [0, 1]$	-	1	Similar to availability, but given as factor of capacity instead. If, e.g., capacity: 100 out:electricity and availability_factor: 0.7, the available capacity will only be 70 electricity. This is especially useful for intermittent generators, where the availability is not a fixed value, but depends on the weather, and can be passed, e.g., by setting availability_factor: wind@input_data_file.

adapt_min_to_availability	no	true, false	-	false	If true, the minimal partial load will be influenced by the availability. Example: Consider a Unit with capacity: 100 out:electricity, a min_conversion of 0.4, and an availability_factor of 0.5. This entails having 50 electricity available, while the minimal partial load is 40 electricity. This results in the Unit at best operating only closely above the minimal partial load. Furthermore, an availability_factor below 0.4 would result in no feasible generation, besides shutting the Unit off. While this might be the intended mode of operation in many use cases, adapt_min_to_availability can change this: If set to true, this dynamically changes the minimal partial load. In the previous example, that means (100 * 0.5) * 0.4 = 20 electricity (the 50% minimum load are now based on the available 40), changing the overall behaviour (including efficiencies) as well as leading to feasible generations even when the availability_factor is below 0.4.

marginal_cost	no	value per dir:carrier	monetary per energy	0	Marginal cost of the consumption/generation of one unit of energy of the specified carrier. Has to be given in the format value per dir:carrier, e.g. 3.5 per out:electricity for a marginal cost of 3.5 monetary units per unit of electricity generated.

enable_ramp_up	no	true, false	-	false	Enables calculation of upward ramps. Ramping is based on the carrier specified in capacity.

enable_ramp_down	no	true, false	-	false	Enables calculation of downward ramps. Ramping is based on the carrier specified in capacity.

ramp_up_cost	no	numeric	monetary per power	0	Sets the cost of ramping up (increasing in-/output) by 1 unit of the capacity carrier.

ramp_down_cost	no	numeric	monetary per power	0	Sets the cost of ramping down (decreasing in-/output) by 1 unit of the capacity carrier.

ramp_up_limit	no	$\in [0, 1]$	-	1	Limits the allowed ramping up based on this factor of the total capacity. If capacity: 100 in:electricity with ramp_up_limit: 0.2, this limits the total increase of usage of electricity (on the input) to 20 units (power) per hour. For example, starting at an input of 35, after one hour the input has to be lesser than or equal to 55. If a Snapshot's duration is set to, e.g., two hours, this would allow a total increase of 40 units.

ramp_down_limit	no	$\in [0, 1]$	-	1	Limits the allowed ramping down based on this factor of the total capacity. See ramp_up_limit.

min_on_time	no	numeric	hours	0	Minimum on-time of the Unit. If set, the Unit has to be on for at least this amount of time, after turning on. It is highly recommended to only use this with unit_commitment: binary, unless you know why it's fine to use with another mode.

min_off_time	no	numeric	hours	0	Minimum off-time of the Unit. If set, the Unit has to be off for at least this amount of time, after turning off. It is highly recommended to only use this with unit_commitment: binary, unless you know why it's fine to use with another mode.

on_time_before	no	numeric	hours	0	Time that this Unit has already been running before the optimization starts. Can be used in combination with min_on_time.

off_time_before	no	numeric	hours	0	Time that this Unit has already been off before the optimization starts. Can be used in combination with min_off_time.

is_on_before	no	numeric	-	1	Number of Units that should be considered to have been running before the optimization starts. Can be used in combination with on_time_before, especially for unit_count greater than 1.

unit_commitment	no	off, linear, binary, integer	-	off	Controls how the unit commitment of this Unit is handled. linear results in the ability to startup parts of the unit (so 0.314159 is a feasible amount of "turned on unit"), while binary restricts the Unit to either be on (converting the conversion_at_min + possible additional conversion above that minimum) or off (converting nothing); integer is needed to consider binary unit commitment for Units with more than 1 "grouped unit" (see unit_count).

unit_count	no	numeric	-	1	Number of units aggregated in this Unit. Besides interacting with the mode of unit_commitment, this mainly is responsible for scaling the output (e.g. grouping 47 of the same wind turbine, ...).

min_conversion	no	$\in [0, 1]$	-	-	If unit_commitment is not set to off, this specifies the percentage that is considered to be the minimal feasible partial load this Unit can operate at. Operating below that setpoint is not allowed, at that point the conversion_at_min coefficients are used, and above that they are scaled to result in conversion when running at full capacity.

conversion_at_min	no	string	-	-	The conversion expression while running on the minimal partial load. Only applicable if unit_commitment is not off and min_conversion is explicitly set. Follows the same form as conversion.

startup_cost	no	numeric	monetary per start	0	Costs per startup (also applicable if startups are not binary or integer). This is necessary to allow conversion_at_min to have (at least partially) the effect that one expects, if unit_commitment: linear.

build_priority	no	numeric	-	0	Priority for the build order of components. Components with higher build_priority are built before. This can be useful for addons, that connect multiple components and rely on specific components being initialized before others.

Detailed Model Reference

Variables

conversion

nothing

ison

nothing

ramp

nothing

startup

nothing

Expressions

Constraints

conversion_bounds

nothing

ison

nothing

min_onoff_time

nothing

ramp

nothing

ramp_limit

nothing

startup

nothing

Objectives

marginal_cost

nothing

ramp_cost

nothing

startup_cost

nothing